fix(editor): reduce recursive action dispatcher stack usage
