+++ src/editor.rs
@@ -19,6 +19,7 @@ mod agent_manager;
 mod buffer_manager;
 mod diagnostics_picker;
 mod display_layout;
+mod inline_actions;
 mod inline_agent_outcomes;
 mod inline_changes;
 mod inline_comment_followup;
@@ -35,6 +36,7 @@ mod lsp_coordinator;
 #[cfg(test)]
 mod navigation_perf_tests;
 mod notifications;
+mod package_actions;
 pub(crate) mod perf;
 pub mod render_buffer;
 pub mod rendering;
@@ -17124,778 +17126,51 @@ impl Editor {
                     self.draw_commandline(buffer);
                 }
             }
-            Action::AddSampleInlineComment => {
+            Action::AddSampleInlineComment
+            | Action::ClearInlineComments
+            | Action::ShowInlineComment
+            | Action::DismissInlineComment
+            | Action::NextInlineComment
+            | Action::PreviousInlineComment
+            | Action::NextOverlappingInlineComment
+            | Action::PreviousOverlappingInlineComment
+            | Action::OpenInlineComment(..)
+            | Action::FocusInlineComment(..)
+            | Action::ChooseInlineComment(..)
+            | Action::NavigateInlineCommentCard { .. }
+            | Action::RefineInlineComment(..)
+            | Action::ResolveInlineComment(..)
+            | Action::DismissInlineCommentById(..)
+            | Action::NavigateOverlappingInlineComment { .. }
+            | Action::AskInlineComment { .. }
+            | Action::InlineAssist
+            | Action::SubmitInlineAssist(..)
+            | Action::HideInlineAssist
+            | Action::SaveInlineAssistDraft
+            | Action::DiscardInlineAssistDraft
+            | Action::OpenInlineJob(..)
+            | Action::OpenLatestInlineCompletion
+            | Action::OpenInlineCompletion(..)
+            | Action::ApplyPendingInlineAssist
+            | Action::ApplyReviewedInlineAssist(..)
+            | Action::ViewInlineChanges { .. }
+            | Action::ViewInlineAgentChanges { .. }
+            | Action::UndoInlineChange(..)
+            | Action::RejectPendingInlineAssist
+            | Action::CancelInlineAssist
+            | Action::CancelInlineAssistRefine
+            | Action::KeepInlineAssist
+            | Action::UndoInlineAssist
+            | Action::RefineInlineAssist
+            | Action::ViewInlineAssistAnswer
+            | Action::EscalateInlineAssist
+            | Action::StageInlineAssistHandoff { .. } => {
                 add_to_history = false;
-                if matches!(self.mode, Mode::Normal | Mode::Visual | Mode::VisualLine) {
-                    let (start_line, end_line) = self.inline_comment_target_lines();
-                    let was_visual = self.is_visual();
-                    self.add_sample_inline_comment();
-                    if was_visual {
-                        self.execute(&Action::EnterMode(Mode::Normal), buffer, runtime)
-                            .await?;
-                        self.move_to_text_position(TextPosition::new(start_line, 0));
-                        self.refresh_cursor_goal();
-                    }
-                    let scope = if start_line == end_line {
-                        format!("line {}", start_line + 1)
-                    } else {
-                        format!("lines {}–{}", start_line + 1, end_line + 1)
-                    };
-                    self.set_legacy_message(Some(format!(
-                        "sample comment · {scope} · Space C replace · Space X clear"
-                    )));
-                    self.render(buffer)?;
-                }
-            }
-            Action::ClearInlineComments => {
-                add_to_history = false;
-                self.clear_inline_comments();
-                self.render(buffer)?;
-            }
-            Action::ShowInlineComment => {
-                add_to_history = false;
-                if let Some(id) = self.current_inline_comment_id() {
-                    self.open_inline_comment_by_id(id, buffer, runtime).await?;
-                } else {
-                    self.show_inline_comment();
-                }
-                self.render(buffer)?;
-            }
-            Action::DismissInlineComment => {
-                add_to_history = false;
-                self.dismiss_inline_comment();
-                self.render(buffer)?;
-            }
-            Action::NextInlineComment | Action::PreviousInlineComment => {
-                add_to_history = false;
-                self.navigate_inline_comment(matches!(action, Action::PreviousInlineComment));
-                self.render(buffer)?;
-            }
-            Action::NextOverlappingInlineComment | Action::PreviousOverlappingInlineComment => {
-                add_to_history = false;
-                if let Some(id) = self.current_inline_comment_id() {
-                    self.cycle_overlapping_inline_comment(
-                        id,
-                        matches!(action, Action::PreviousOverlappingInlineComment),
-                    );
-                } else {
-                    self.set_legacy_message(Some("no inline item at the cursor".into()));
-                }
-                self.render(buffer)?;
-            }
-            Action::OpenInlineComment(id) => {
-                add_to_history = false;
-                self.open_inline_comment_by_id(*id, buffer, runtime).await?;
-            }
-            Action::FocusInlineComment(id) => {
-                add_to_history = false;
-                self.focus_inline_comment(*id, buffer, runtime).await?;
-            }
-            Action::ChooseInlineComment(id) => {
-                add_to_history = false;
-                self.choose_inline_comment(*id);
-                self.render(buffer)?;
-            }
-            Action::NavigateInlineCommentCard { id, backwards } => {
-                add_to_history = false;
-                if let Some(next) = self.cycle_overlapping_inline_comment(*id, *backwards) {
-                    self.focus_inline_comment(next, buffer, runtime).await?;
-                }
-            }
-            Action::RefineInlineComment(id) => {
-                add_to_history = false;
-                self.refine_inline_comment(*id, buffer, runtime).await?;
-            }
-            Action::ResolveInlineComment(id) => {
-                add_to_history = false;
-                self.resolve_inline_comment(*id);
-                self.render(buffer)?;
-            }
-            Action::DismissInlineCommentById(id) => {
-                add_to_history = false;
-                if self.select_inline_comment_by_id(*id) {
-                    self.dismiss_inline_comment();
-                    self.current_dialog = None;
-                }
-                self.render(buffer)?;
-            }
-            Action::NavigateOverlappingInlineComment {
-                id,
-                backwards,
-                open,
-            } => {
-                add_to_history = false;
-                if *open {
-                    self.park_inline_assist();
-                }
-                if let Some(selected) = self.cycle_overlapping_inline_comment(*id, *backwards) {
-                    if *open {
-                        self.open_inline_comment_by_id(selected, buffer, runtime)
-                            .await?;
-                    } else {
-                        self.render(buffer)?;
-                    }
-                }
-            }
-            Action::AskInlineComment { id, in_agent } => {
-                add_to_history = false;
-                self.ask_inline_comment(*id, *in_agent, buffer, runtime)
-                    .await?;
-            }
-            Action::InlineAssist => {
-                add_to_history = false;
-                match self.inline_assist_target() {
-                    Ok((range, scope)) => {
-                        let Some(window_id) = self.window_manager.active_stable_window_id() else {
-                            self.set_legacy_message(Some(
-                                "inline assist requires an active editor window".to_string(),
-                            ));
-                            self.draw_commandline(buffer);
-                            return Ok(false);
-                        };
-                        let expected_text = self.current_buffer().text_in_range(range);
-                        self.park_inline_assist();
-                        self.inline_assist = Some(InlineAssistSession {
-                            parent_comment: None,
-                            allow_expansion: !matches!(
-                                self.mode,
-                                Mode::Visual | Mode::VisualLine | Mode::VisualBlock
-                            ),
-                            buffer_id: self.current_buffer().id(),
-                            window_id,
-                            expected_revision: self.current_buffer().revision(),
-                            range,
-                            expected_text,
-                            scope: scope.clone(),
-                            request_id: None,
-                            session_id: None,
-                            transaction_id: None,
-                            annotation_group_id: uuid::Uuid::new_v4().to_string(),
-                            has_result: false,
-                            result_request_id: None,
-                        });
-                        self.current_dialog = Some(Box::new(self.inline_assist_popup(
-                            scope,
-                            InlineAssistPopupState::Prompt {
-                                initial: String::new(),
-                                refining: false,
-                            },
-                        )));
-                        self.render(buffer)?;
-                    }
-                    Err(error) => {
-                        self.set_legacy_message(Some(error.to_string()));
-                        self.draw_commandline(buffer);
-                    }
-                }
-            }
-            Action::SubmitInlineAssist(prompt) => {
-                add_to_history = false;
-                let range = match self.inline_submission_target() {
-                    Ok(range) => range,
-                    Err(error) => {
-                        self.set_legacy_message(Some(error.to_string()));
-                        self.render(buffer)?;
-                        return Ok(false);
-                    }
-                };
-                let Some((scope, existing_session, retired_session)) =
-                    self.inline_assist.as_ref().map(|assist| {
-                        let reuse = range == assist.range && self.inline_target_matches(assist);
-                        (
-                            assist.scope.clone(),
-                            reuse.then(|| assist.session_id.clone()).flatten(),
-                            (!reuse).then(|| assist.session_id.clone()).flatten(),
-                        )
-                    })
-                else {
-                    self.set_legacy_message(Some("inline assist is no longer active".to_string()));
-                    return Ok(false);
-                };
-                let mut context = match self.inline_assist_context(range) {
-                    Ok(context) => context,
-                    Err(error) => {
-                        self.set_legacy_message(Some(error.to_string()));
-                        return Ok(false);
-                    }
-                };
-                if existing_session.is_none() {
-                    if let Some(assist) = &self.inline_assist {
-                        context
-                            .push_str(&self.recovered_inline_context(&assist.annotation_group_id));
-                    }
-                }
-                let request_id = uuid::Uuid::new_v4().to_string();
-                if let Err(error) = self.begin_inline_history_turn(&request_id, prompt, range) {
-                    self.set_legacy_message(Some(error.to_string()));
-                    self.render(buffer)?;
-                    return Ok(false);
-                }
-                if let Some(session_id) = retired_session {
-                    if let Some(bridge) = self.agent_manager.bridge() {
-                        let _ = bridge.try_send(CodexCommand::CloseSession { session_id });
-                    }
-                }
-                let revision = self.current_buffer().revision();
-                let expected_text = self.current_buffer().text_in_range(range);
-                let buffer_id = self.current_buffer().id();
-                if let Some(assist) = self.inline_assist.as_mut() {
-                    assist.buffer_id = buffer_id;
-                    assist.range = range;
-                    assist.expected_revision = revision;
-                    assist.expected_text = expected_text;
-                    assist.request_id = Some(request_id.clone());
-                    assist.session_id.clone_from(&existing_session);
-                }
-                let cwd = get_workspace_path();
-                if let Err(error) = self.ensure_agent_bridge(&cwd) {
-                    self.inline_history.finish(
-                        &request_id,
-                        InlineTurnState::Failed,
-                        Some(error.to_string()),
-                    );
-                    self.current_dialog = Some(Box::new(self.inline_assist_popup(
-                        scope,
-                        InlineAssistPopupState::Failed(error.to_string()),
-                    )));
-                    self.sync_inline_activity();
-                    self.render(buffer)?;
-                    return Ok(false);
-                }
-                let command = existing_session.map_or_else(
-                    || CodexCommand::InlineAssist {
-                        request_id: request_id.clone(),
-                        cwd,
-                        prompt: prompt.clone(),
-                        context: context.clone(),
-                    },
-                    |session_id| CodexCommand::InlineAssistFollowup {
-                        request_id: request_id.clone(),
-                        session_id,
-                        prompt: prompt.clone(),
-                        context: context.clone(),
-                    },
-                );
-                let send_result = if let Some(bridge) = self.agent_manager.bridge() {
-                    bridge.send(command).await
-                } else {
-                    Err(anyhow::anyhow!("Codex bridge is unavailable"))
-                };
-                if let Err(error) = send_result {
-                    self.inline_history.finish(
-                        &request_id,
-                        InlineTurnState::Failed,
-                        Some(error.to_string()),
-                    );
-                    if let Some(assist) = self.inline_assist.as_mut() {
-                        assist.session_id = None;
-                    }
-                    self.current_dialog = Some(Box::new(self.inline_assist_popup(
-                        scope,
-                        InlineAssistPopupState::Failed(error.to_string()),
-                    )));
-                } else {
-                    self.current_dialog = Some(Box::new(
-                        self.inline_assist_popup(scope, InlineAssistPopupState::Working),
-                    ));
-                }
-                self.sync_inline_activity();
-                self.render(buffer)?;
-            }
-            Action::HideInlineAssist => {
-                add_to_history = false;
-                let close = self
-                    .current_dialog
-                    .as_mut()
-                    .and_then(|dialog| dialog.request_inline_assist_close());
-                if let Some(close) = close {
-                    self.execute_with_tracking(&close, buffer, runtime, false)
-                        .await?;
-                } else {
-                    self.park_inline_assist();
-                    self.render(buffer)?;
-                }
-            }
-            Action::SaveInlineAssistDraft => {
-                add_to_history = false;
-                if self.current_dialog.as_ref().is_some_and(|dialog| {
-                    matches!(
-                        dialog.inline_assist_state(),
-                        Some(InlineAssistPopupState::Prompt { .. })
-                    )
-                }) {
-                    self.park_inline_assist();
-                }
-                self.render(buffer)?;
-            }
-            Action::DiscardInlineAssistDraft => {
-                add_to_history = false;
-                if self.current_dialog.as_ref().is_some_and(|dialog| {
-                    matches!(
-                        dialog.inline_assist_state(),
-                        Some(InlineAssistPopupState::Prompt { .. })
-                    )
-                }) {
-                    self.close_inline_assist_session();
-                }
-                self.render(buffer)?;
-            }
-            Action::OpenInlineJob(group) => {
-                add_to_history = false;
-                self.open_inline_job(group, buffer, runtime).await?;
-            }
-            Action::OpenLatestInlineCompletion => {
-                add_to_history = false;
-                self.open_latest_inline_completion(buffer, runtime).await?;
-            }
-            Action::OpenInlineCompletion(request) => {
-                add_to_history = false;
-                self.open_inline_completion(request, buffer, runtime)
-                    .await?;
-            }
-            Action::ApplyPendingInlineAssist | Action::ApplyReviewedInlineAssist(_) => {
-                add_to_history = false;
-                let pending = self
-                    .inline_assist
-                    .as_ref()
-                    .and_then(|assist| assist.request_id.as_deref())
-                    .and_then(|request| self.inline_history.turn(request))
-                    .filter(|turn| turn.state == InlineTurnState::Ready)
-                    .and_then(|turn| {
-                        Some((
-                            turn.request_id.clone(),
-                            turn.session_id.clone()?,
-                            turn.result.clone()?,
-                        ))
-                    });
-                if let Some((request, session, result)) = pending {
-                    if let Action::ApplyReviewedInlineAssist(reviewed) = action {
-                        if reviewed != &request {
-                            self.set_legacy_message(Some(
-                                "inline edit review is no longer current".into(),
-                            ));
-                            self.render(buffer)?;
-                            return Ok(false);
-                        }
-                    } else if result.expanded_scope.is_some() {
-                        self.execute_with_tracking(
-                            &Action::ViewInlineAssistAnswer,
-                            buffer,
-                            runtime,
-                            false,
-                        )
-                        .await?;
-                        return Ok(false);
-                    }
-                    let preparation = if result.expanded_scope.is_some() {
-                        self.prepare_reviewed_inline_expansion(&request)
-                    } else {
-                        Ok(())
-                    };
-                    let applied = match preparation {
-                        Ok(()) => {
-                            self.apply_inline_result(&request, &session, &result, buffer, runtime)
-                                .await
-                        }
-                        Err(error) => Err(error),
-                    };
-                    if let Err(error) = applied {
-                        self.inline_history.finish(
-                            &request,
-                            InlineTurnState::Rejected,
-                            Some(error.to_string()),
-                        );
-                        self.sync_inline_activity();
-                        let scope = self
-                            .inline_assist
-                            .as_ref()
-                            .map_or("selection", |assist| assist.scope.as_str())
-                            .to_string();
-                        self.current_dialog = Some(Box::new(self.inline_assist_popup(
-                            scope,
-                            InlineAssistPopupState::Failed(error.to_string()),
-                        )));
-                        self.render(buffer)?;
-                    }
-                }
-            }
-            Action::ViewInlineChanges { request_id, hunk } => {
-                add_to_history = false;
-                self.view_inline_changes(request_id, *hunk, buffer, runtime)
-                    .await?;
-            }
-            Action::ViewInlineAgentChanges {
-                request_id,
-                outcome,
-                change,
-            } => {
-                add_to_history = false;
-                self.view_inline_agent_changes(request_id, *outcome, *change, buffer, runtime)
-                    .await?;
-            }
-            Action::UndoInlineChange(request) => {
-                add_to_history = false;
-                self.undo_inline_change(request, buffer, runtime).await?;
-            }
-            Action::RejectPendingInlineAssist => {
-                add_to_history = false;
-                if let Some(request) = self
-                    .inline_assist
-                    .as_ref()
-                    .and_then(|assist| assist.request_id.clone())
-                    .filter(|request| {
-                        self.inline_history.turn(request).is_some_and(|turn| {
-                            turn.state == InlineTurnState::Ready
-                                && turn
-                                    .result
-                                    .as_ref()
-                                    .is_some_and(|result| result.changes_text(&turn.before))
-                        })
-                    })
+                if let std::ops::ControlFlow::Break(quit) =
+                    self.execute_inline_action(action, buffer, runtime).await?
                 {
-                    self.inline_history.finish(
-                        &request,
-                        InlineTurnState::Declined,
-                        Some("Inline edit declined; source unchanged.".into()),
-                    );
-                    self.close_inline_assist_session();
-                    self.notify_inline_outcome(&request);
-                    self.set_legacy_message(Some(
-                        "inline edit declined · source unchanged · retained in InlineHistory"
-                            .into(),
-                    ));
-                    self.render(buffer)?;
+                    return Ok(quit);
                 }
-            }
-            Action::CancelInlineAssist => {
-                add_to_history = false;
-                self.close_inline_assist_session();
-                self.render(buffer)?;
-            }
-            Action::CancelInlineAssistRefine => {
-                add_to_history = false;
-                if let Some(scope) = self
-                    .inline_assist
-                    .as_ref()
-                    .map(|assist| assist.scope.clone())
-                {
-                    self.current_dialog = Some(Box::new(
-                        self.inline_assist_popup(scope, self.inline_assist_result_state()),
-                    ));
-                } else {
-                    self.current_dialog = None;
-                }
-                self.render(buffer)?;
-            }
-            Action::KeepInlineAssist => {
-                add_to_history = false;
-                let result = self.inline_assist_result_state();
-                let request = self
-                    .inline_assist
-                    .as_ref()
-                    .and_then(|assist| assist.result_request_id.clone());
-                self.close_inline_assist_session();
-                if matches!(result, InlineAssistPopupState::Applied { edited: true, .. }) {
-                    self.set_legacy_message(None);
-                    if let Some(request) = request {
-                        self.notify_inline_outcome(&request);
-                    }
-                } else {
-                    self.set_legacy_message(Some(match result {
-                        InlineAssistPopupState::Applied { comments, .. } => format!(
-                            "kept {comments} inline comment(s) · Space v view · Space x dismiss"
-                        ),
-                        InlineAssistPopupState::NeedsAgent(_) => {
-                            "discussion kept · Space H to continue from history".to_string()
-                        }
-                        _ => "inline assist closed".to_string(),
-                    }));
-                }
-                self.render(buffer)?;
-            }
-            Action::UndoInlineAssist => {
-                add_to_history = false;
-                let transaction_id = self
-                    .inline_assist
-                    .as_ref()
-                    .and_then(|assist| assist.transaction_id.clone());
-                let is_latest = transaction_id.as_deref().is_some_and(|transaction_id| {
-                    self.current_buffer()
-                        .undo_history
-                        .latest_transaction()
-                        .is_some_and(|latest| latest.id == transaction_id)
-                });
-                if let Some(request) = self
-                    .inline_assist
-                    .as_ref()
-                    .and_then(|assist| assist.result_request_id.as_deref())
-                {
-                    if let Some(turn) = self.inline_history.turn_mut(request) {
-                        if is_latest || transaction_id.is_none() {
-                            turn.disposition = InlineDisposition::Undone;
-                        }
-                        if let Some(result) = &turn.result {
-                            turn.hidden_comments = (0..result.comments.len()).collect();
-                        }
-                    }
-                }
-                if is_latest {
-                    for turn in self
-                        .inline_history
-                        .conversations
-                        .iter_mut()
-                        .flat_map(|conversation| &mut conversation.turns)
-                    {
-                        if turn.transaction_id.as_deref() == transaction_id.as_deref() {
-                            turn.disposition = InlineDisposition::Undone;
-                        }
-                    }
-                }
-                if let Some(group_id) = self
-                    .inline_assist
-                    .as_ref()
-                    .map(|assist| assist.annotation_group_id.clone())
-                {
-                    self.dismiss_inline_comment_group(&group_id);
-                }
-                self.close_inline_assist_session();
-                if is_latest {
-                    self.undo_transaction(buffer, runtime).await?;
-                } else if transaction_id.is_some() {
-                    self.set_legacy_message(Some(
-                        "inline edit is no longer the latest change; use transaction history to revert it"
-                            .to_string(),
-                    ));
-                    self.render(buffer)?;
-                } else {
-                    self.set_legacy_message(Some("dismissed inline comments".to_string()));
-                    self.render(buffer)?;
-                }
-            }
-            Action::RefineInlineAssist => {
-                add_to_history = false;
-                if let Err(error) = self.inline_submission_target() {
-                    self.set_legacy_message(Some(error.to_string()));
-                    self.render(buffer)?;
-                    return Ok(false);
-                }
-                let initial = self
-                    .inline_assist
-                    .as_ref()
-                    .and_then(|assist| assist.request_id.as_deref())
-                    .and_then(|request| self.inline_history.turn(request))
-                    .filter(|turn| turn.state == InlineTurnState::Ready)
-                    .map_or_else(String::new, |turn| {
-                        format!(
-                            "Recheck this request against the current code: {}",
-                            turn.prompt
-                        )
-                    });
-                if let Some((scope, refining)) = self
-                    .inline_assist
-                    .as_ref()
-                    .map(|assist| (assist.scope.clone(), assist.has_result))
-                {
-                    self.current_dialog = Some(Box::new(self.inline_assist_popup(
-                        scope,
-                        InlineAssistPopupState::Prompt { initial, refining },
-                    )));
-                    self.render(buffer)?;
-                }
-            }
-            Action::ViewInlineAssistAnswer => {
-                add_to_history = false;
-                if let Some(request) = self
-                    .inline_assist
-                    .as_ref()
-                    .and_then(|assist| assist.request_id.as_deref())
-                    .and_then(|request| self.inline_history.turn(request))
-                    .filter(|turn| turn.has_code_change())
-                    .map(|turn| turn.request_id.clone())
-                {
-                    self.view_inline_changes(&request, 0, buffer, runtime)
-                        .await?;
-                    return Ok(false);
-                }
-                if let Some(turn) = self
-                    .inline_assist
-                    .as_ref()
-                    .and_then(|assist| {
-                        assist
-                            .request_id
-                            .as_deref()
-                            .or(assist.result_request_id.as_deref())
-                    })
-                    .and_then(|request| self.inline_history.turn(request))
-                {
-                    let expanded_range = self
-                        .inline_assist
-                        .as_ref()
-                        .and_then(|session| self.pending_inline_expansion_range(session).ok());
-                    let proposed = turn.proposed_edit();
-                    let description = if proposed.is_some() {
-                        turn.proposal_description()
-                    } else {
-                        turn.answer_text()
-                    };
-                    let answer = expanded_range.map_or_else(
-                        || description.clone(),
-                        |range| {
-                            format!(
-                                "Current proposed range: {}:{}–{}\n\n{}",
-                                turn.location.file,
-                                range.start.line + 1,
-                                range.end.line + usize::from(range.end.character > 0),
-                                description
-                            )
-                        },
-                    );
-                    let mut hover =
-                        HoverInfo::new(self, answer, HoverInfoFormat::Markdown, Vec::new())
-                            .with_label("Inline answer")
-                            .with_close_action(Action::CancelInlineAssistRefine);
-                    if let Some((before, after)) = proposed {
-                        hover = hover.with_diff(&turn.location.file, before, after, "proposed");
-                    }
-                    if turn.state == InlineTurnState::Ready
-                        && turn
-                            .result
-                            .as_ref()
-                            .is_some_and(|result| result.expanded_scope.is_some())
-                    {
-                        hover = hover.with_label("Review wider edit · not applied");
-                        if expanded_range.is_some() {
-                            hover = hover.with_shortcut(
-                                'a',
-                                "apply wider edit",
-                                Action::ApplyReviewedInlineAssist(turn.request_id.clone()),
-                            );
-                        }
-                        hover =
-                            hover.with_shortcut('d', "decline", Action::RejectPendingInlineAssist);
-                    } else if turn.state == InlineTurnState::Ready
-                        && turn
-                            .result
-                            .as_ref()
-                            .is_some_and(|result| result.changes_text(&turn.before))
-                    {
-                        hover = hover
-                            .with_label("Review inline edit · not applied")
-                            .with_shortcut('d', "decline", Action::RejectPendingInlineAssist);
-                        if self
-                            .inline_assist
-                            .as_ref()
-                            .is_some_and(|session| self.inline_target_matches(session))
-                        {
-                            hover = hover.with_shortcut(
-                                'a',
-                                "apply edit",
-                                Action::ApplyReviewedInlineAssist(turn.request_id.clone()),
-                            );
-                        }
-                    } else if let Some((id, ordinal, count)) = self.current_inline_navigation() {
-                        hover = hover
-                            .with_label(format!("Inline answer · inline {ordinal} of {count}"))
-                            .with_inline_navigation(id);
-                    }
-                    self.current_dialog = Some(Box::new(hover));
-                }
-                self.render(buffer)?;
-            }
-            Action::EscalateInlineAssist => {
-                add_to_history = false;
-                let Some(assist) = self.inline_assist.as_ref() else {
-                    return Ok(false);
-                };
-                let Some(prompt) = self.inline_handoff_prompt(&assist.annotation_group_id) else {
-                    self.set_legacy_message(Some(
-                        "inline discussion is no longer available".into(),
-                    ));
-                    return Ok(false);
-                };
-                let request_id = assist.request_id.clone();
-                self.plugin_registry
-                    .ensure_command_registered(runtime, "AgentOpen")
-                    .await;
-                if runtime.command_plugin("AgentOpen").is_none() {
-                    self.set_legacy_message(Some(
-                        "Agent is unavailable; the inline discussion remains in history".into(),
-                    ));
-                    self.render(buffer)?;
-                    return Ok(false);
-                }
-                self.close_inline_assist_session();
-                // The handoff carries its own location and source context. Do not
-                // leave an artificial visual selection behind in the editor.
-                if !self.is_normal() {
-                    self.execute_with_tracking(
-                        &Action::EnterMode(Mode::Normal),
-                        buffer,
-                        runtime,
-                        false,
-                    )
-                    .await?;
-                }
-                self.plugin_registry.execute(runtime, "AgentOpen").await?;
-                // AgentOpen queues pane creation/restoration first. Stage the draft
-                // only after those requests have been processed.
-                runtime.send_request(PluginRequest::Action(Action::StageInlineAssistHandoff {
-                    request_id,
-                    comment_followup: None,
-                    prompt,
-                    expected_draft: None,
-                }));
-            }
-            Action::StageInlineAssistHandoff {
-                request_id,
-                comment_followup,
-                prompt,
-                expected_draft,
-            } => {
-                add_to_history = false;
-                // Session restoration can hide a newly recreated pane, and an
-                // editor zoom can obscure an otherwise visible one. An explicit
-                // handoff must reveal the actual pane before loading its draft.
-                if self
-                    .panel_manager
-                    .set_panel_visible("agent-conversation", true)
-                {
-                    self.clear_pane_zoom();
-                    self.apply_panel_layout();
-                }
-                match self.panel_manager.load_text_panel_draft(
-                    "agent-conversation",
-                    prompt,
-                    *expected_draft,
-                ) {
-                    Ok(plugin::panel::TextPanelReuseOutcome::Loaded) => {
-                        self.staged_inline_agent_handoff = request_id
-                            .as_ref()
-                            .filter(|request| {
-                                comment_followup.is_some()
-                                    || self.inline_history.turn(request).is_some()
-                            })
-                            .map(|request| inline_agent_outcomes::StagedHandoff {
-                                request_id: request.clone(),
-                                comment_followup: comment_followup.clone(),
-                            });
-                        self.set_legacy_message(Some(
-                            "inline discussion loaded in Agent; review and send when ready".into(),
-                        ));
-                    }
-                    Ok(plugin::panel::TextPanelReuseOutcome::Confirm(revision)) => {
-                        self.current_dialog = Some(Box::new(Confirmation::new_actions(
-                            self,
-                            "Replace unsent Agent draft?",
-                            "The inline discussion will replace the draft as one undoable edit. Nothing will be sent.",
-                            "Load discussion", "Keep draft",
-                            Action::StageInlineAssistHandoff { request_id: request_id.clone(), comment_followup: comment_followup.clone(), prompt: prompt.clone(), expected_draft: Some(revision) },
-                            Action::Print("current draft kept; inline discussion remains in history".into()),
-                        )));
-                    }
-                    Err(message) => self.set_legacy_message(Some(message.to_string())),
-                }
-                self.render(buffer)?;
             }
             Action::EnterMode(new_mode) => {
                 add_to_history = false;
@@ -19044,529 +18319,18 @@ impl Editor {
                     self.set_legacy_message(Some("No log file configured".to_string()));
                 }
             }
-            Action::ListPlugins => {
+            Action::ListPlugins
+            | Action::PluginManagerCatalogLoaded { .. }
+            | Action::PluginManagerSelect(..)
+            | Action::PluginManagerCatalogConsent { .. }
+            | Action::PluginManagerTrustConsent(..)
+            | Action::PluginManagerTrustConfirmed { .. }
+            | Action::PluginManagerCatalogInstall { .. }
+            | Action::PluginManagerInstall(..)
+            | Action::PluginManagerAction(..)
+            | Action::PluginManagerFinished { .. } => {
                 add_to_history = false;
-                let manager = plugin::package::PluginPackageManager::new(Config::config_dir());
-                let installed = match manager.list() {
-                    Ok(installed) => installed,
-                    Err(error) => {
-                        self.set_legacy_message(Some(error.to_string()));
-                        Vec::new()
-                    }
-                };
-                let items = plugin_manager_items(
-                    &installed,
-                    &self.plugin_catalog,
-                    &self.plugin_catalog_url,
-                );
-                let picker = language_pack_picker(self, items, "Loading official catalog…", true);
-                self.current_dialog = Some(Box::new(picker));
-                let catalog_url = plugin::catalog::catalog_url();
-                let runtime = runtime.clone();
-                tokio::spawn(async move {
-                    let (packages, error) =
-                        match plugin::catalog::PluginCatalog::fetch(&catalog_url).await {
-                            Ok(catalog) => (catalog.packages, None),
-                            Err(error) => (Vec::new(), Some(error.to_string())),
-                        };
-                    runtime.send_request(PluginRequest::Action(
-                        Action::PluginManagerCatalogLoaded {
-                            catalog_url,
-                            packages,
-                            error,
-                        },
-                    ));
-                });
-                self.render(buffer)?;
-            }
-            Action::PluginManagerCatalogLoaded {
-                catalog_url,
-                packages,
-                error,
-            } => {
-                add_to_history = false;
-                if error.is_none() {
-                    self.plugin_catalog = packages
-                        .iter()
-                        .cloned()
-                        .map(|package| (package.id.clone(), package))
-                        .collect();
-                    self.plugin_catalog_url = catalog_url.clone();
-                }
-                let manager = plugin::package::PluginPackageManager::new(Config::config_dir());
-                let installed = manager.list().unwrap_or_default();
-                let mut items = plugin_manager_items(
-                    &installed,
-                    &self.plugin_catalog,
-                    &self.plugin_catalog_url,
-                );
-                if error.is_some() {
-                    items.insert(
-                        0,
-                        plugin_manager_action_item(
-                            "retry-catalog".to_string(),
-                            "Retry official catalog",
-                            "The last refresh failed",
-                            "Retry",
-                        ),
-                    );
-                }
-                let initial_catalog_selection = if error.is_some() {
-                    Some("retry-catalog".to_string())
-                } else if installed.is_empty() {
-                    items
-                        .iter()
-                        .find(|item| item.id.starts_with("catalog:"))
-                        .map(|item| item.id.clone())
-                } else {
-                    None
-                };
-                if let Some(dialog) = &mut self.current_dialog {
-                    dialog.update_picker(PLUGIN_MANAGER_PICKER_ID, PickerUpdate::Items(items));
-                    if let Some(selection) = initial_catalog_selection {
-                        dialog.update_picker(
-                            PLUGIN_MANAGER_PICKER_ID,
-                            PickerUpdate::Selection(selection),
-                        );
-                    }
-                    dialog.update_picker(
-                        PLUGIN_MANAGER_PICKER_ID,
-                        PickerUpdate::Status(Some(error.as_ref().map_or_else(
-                            || {
-                                format!(
-                                    "{} pack{} · Enter open",
-                                    self.plugin_catalog.len(),
-                                    if self.plugin_catalog.len() == 1 {
-                                        ""
-                                    } else {
-                                        "s"
-                                    }
-                                )
-                            },
-                            |error| format!("Catalog unavailable: {error} · Enter to retry"),
-                        ))),
-                    );
-                    dialog.update_picker(PLUGIN_MANAGER_PICKER_ID, PickerUpdate::Busy(false));
-                }
-                self.render(buffer)?;
-            }
-            Action::PluginManagerSelect(selection) => {
-                add_to_history = false;
-                if selection == "retry-catalog" {
-                    self.set_legacy_message(Some(
-                        "Retrying official language-pack catalog…".to_string(),
-                    ));
-                    runtime.send_request(PluginRequest::Action(Action::ListPlugins));
-                } else if selection == "custom-source" {
-                    self.current_dialog = Some(Box::new(InputPrompt::new(
-                        self,
-                        "GitHub owner/repo[@ref] or local path",
-                        "",
-                        Action::PluginManagerInstall,
-                    )));
-                } else if let Some(raw_id) = selection.strip_prefix("unavailable:") {
-                    let id = plugin::package::PluginId::parse(raw_id)?;
-                    let message = self
-                        .plugin_catalog
-                        .get(&id)
-                        .and_then(|package| catalog_package_availability(package).1)
-                        .unwrap_or_else(|| {
-                            format!(
-                                "Language pack `{id}` changed availability; select it again to continue."
-                            )
-                        });
-                    self.set_legacy_message(Some(message.clone()));
-                    let manager = plugin::package::PluginPackageManager::new(Config::config_dir());
-                    let installed = manager.list().unwrap_or_default();
-                    let items = plugin_manager_items(
-                        &installed,
-                        &self.plugin_catalog,
-                        &self.plugin_catalog_url,
-                    );
-                    let picker = language_pack_picker(self, items, message, false);
-                    self.current_dialog = Some(Box::new(picker));
-                } else if let Some(raw_id) = selection.strip_prefix("installed:") {
-                    let id = plugin::package::PluginId::parse(raw_id)?;
-                    let manager = plugin::package::PluginPackageManager::new(Config::config_dir());
-                    let installed = manager
-                        .installed(&id)?
-                        .ok_or_else(|| anyhow::anyhow!("plugin `{id}` is no longer installed"))?;
-                    let toggle = if installed.enabled {
-                        "disable"
-                    } else {
-                        "enable"
-                    };
-                    let mut actions = vec![plugin_manager_action_item(
-                        format!("update\t{id}"),
-                        "Check for updates",
-                        format!("Installed v{}", installed.version),
-                        "Update",
-                    )];
-                    actions.push(plugin_manager_action_item(
-                        format!("{toggle}\t{id}"),
-                        if installed.enabled {
-                            "Disable language pack"
-                        } else {
-                            "Enable language pack"
-                        },
-                        if installed.enabled {
-                            "Preserves files and saved data"
-                        } else {
-                            "Restore language support"
-                        },
-                        if installed.enabled {
-                            "Disable"
-                        } else {
-                            "Enable"
-                        },
-                    ));
-                    if installed.has_native_grammars {
-                        actions.push(plugin_manager_action_item(
-                            format!("trust\t{id}"),
-                            "Approve native grammar",
-                            "Trust the currently installed grammar bytes",
-                            "Warning",
-                        ));
-                    }
-                    actions.push(plugin_manager_action_item(
-                        format!("remove\t{id}"),
-                        "Remove language pack",
-                        "Saved data is preserved",
-                        "Delete",
-                    ));
-                    let status = format!(
-                        "v{} · {} · {}",
-                        installed.version,
-                        if installed.enabled {
-                            "Enabled"
-                        } else {
-                            "Disabled"
-                        },
-                        installed.languages.join(", ")
-                    );
-                    let picker = Picker::builder()
-                        .title(&installed.name)
-                        .structured_items(actions)
-                        .id(PLUGIN_MANAGER_ACTION_PICKER_ID)
-                        .placeholder("Filter actions")
-                        .status(status)
-                        .content_sized(84, 8)
-                        .select_action(|operation| {
-                            if let Some(package) = operation.strip_prefix("trust\t") {
-                                Action::PluginManagerTrustConsent(package.to_string())
-                            } else {
-                                Action::PluginManagerAction(operation)
-                            }
-                        })
-                        .build(self);
-                    self.current_dialog = Some(Box::new(picker));
-                } else if let Some(raw_id) = selection.strip_prefix("catalog:") {
-                    let id = plugin::package::PluginId::parse(raw_id)?;
-                    let manager = plugin::package::PluginPackageManager::new(Config::config_dir());
-                    let replacing_custom = manager.installed(&id)?.is_some_and(|installed| {
-                        !installed_from_catalog(&installed, &self.plugin_catalog_url, &id)
-                    });
-                    let package = self.plugin_catalog.get(&id).ok_or_else(|| {
-                        anyhow::anyhow!("language pack `{id}` is no longer in the catalog")
-                    })?;
-                    anyhow::ensure!(
-                        package.supports_current_red_release()?,
-                        "language pack `{id}` requires Red API `{}`, which this Red release does not support",
-                        package.red_api
-                    );
-                    let artifact = package
-                        .artifact(crate::language::host_target())
-                        .ok_or_else(|| {
-                            anyhow::anyhow!(
-                                "language pack `{id}` is unavailable for `{}`",
-                                crate::language::host_target()
-                            )
-                        })?;
-                    let has_native_grammars = !artifact.grammars.is_empty();
-                    let mut choices = Vec::new();
-                    if has_native_grammars {
-                        choices.push(plugin_manager_action_item(
-                            "install-and-trust".to_string(),
-                            "Install with syntax highlighting",
-                            format!("Approve {} verified grammar(s)", artifact.grammars.len()),
-                            "Warning",
-                        ));
-                    }
-                    choices.push(plugin_manager_action_item(
-                        "install".to_string(),
-                        if has_native_grammars {
-                            "Install without native highlighting"
-                        } else {
-                            "Install language pack"
-                        },
-                        if has_native_grammars {
-                            "Native grammar remains disabled"
-                        } else {
-                            "No native grammar approval required"
-                        },
-                        "Install",
-                    ));
-                    choices.push(plugin_manager_action_item(
-                        "cancel".to_string(),
-                        "Back to language packs",
-                        "Leave the package unchanged",
-                        "Cancel",
-                    ));
-                    let confirmed_package = package.clone();
-                    let confirmed_catalog_url = self.plugin_catalog_url.clone();
-                    let mut status = catalog_package_action_status(package);
-                    if replacing_custom {
-                        status.push_str(" · Replaces custom install");
-                    }
-                    let picker = Picker::builder()
-                        .title(&format!("Install {}", package.name))
-                        .structured_items(choices)
-                        .id(PLUGIN_MANAGER_INSTALL_PICKER_ID)
-                        .placeholder("Filter install options")
-                        .status(status)
-                        .content_sized(84, 8)
-                        .select_action(move |choice| match choice.as_str() {
-                            "cancel" => Action::ListPlugins,
-                            "install-and-trust" => Action::PluginManagerCatalogConsent {
-                                catalog_url: confirmed_catalog_url.clone(),
-                                package: Box::new(confirmed_package.clone()),
-                            },
-                            _ => confirmed_catalog_install_action(
-                                &confirmed_catalog_url,
-                                &confirmed_package,
-                                false,
-                            ),
-                        })
-                        .build(self);
-                    self.current_dialog = Some(Box::new(picker));
-                }
-                self.render(buffer)?;
-            }
-            Action::PluginManagerCatalogConsent {
-                catalog_url,
-                package,
-            } => {
-                add_to_history = false;
-                let message = native_grammar_consent_message(package)?;
-                let accept = confirmed_catalog_install_action(catalog_url, package, true);
-                let cancel = Action::PluginManagerSelect(format!("catalog:{}", package.id));
-                self.current_dialog = Some(Box::new(Confirmation::new_actions(
-                    self,
-                    format!("Approve native grammars for {}", package.name),
-                    message,
-                    "Approve and install",
-                    "Back",
-                    accept,
-                    cancel,
-                )));
-                self.render(buffer)?;
-            }
-            Action::PluginManagerTrustConsent(raw_id) => {
-                add_to_history = false;
-                let id = plugin::package::PluginId::parse(raw_id)?;
-                let manager = plugin::package::PluginPackageManager::new(Config::config_dir());
-                let installed = manager
-                    .installed(&id)?
-                    .ok_or_else(|| anyhow::anyhow!("plugin `{id}` is no longer installed"))?;
-                let digests = manager.package_grammar_digests(&id)?;
-                let message = installed_grammar_consent_message(&installed, &digests)?;
-                let accept = Action::PluginManagerTrustConfirmed {
-                    package: id.to_string(),
-                    digests,
-                };
-                let cancel = Action::PluginManagerSelect(format!("installed:{id}"));
-                self.current_dialog = Some(Box::new(Confirmation::new_actions(
-                    self,
-                    format!("Approve native grammars for {}", installed.name),
-                    message,
-                    "Approve exact bytes",
-                    "Back",
-                    accept,
-                    cancel,
-                )));
-                self.render(buffer)?;
-            }
-            Action::PluginManagerTrustConfirmed { package, digests } => {
-                add_to_history = false;
-                let package = package.clone();
-                let digests = digests.clone();
-                self.set_legacy_message(Some("Approving native grammar bytes…".to_string()));
-                let runtime = runtime.clone();
-                tokio::spawn(async move {
-                    let manager = plugin::package::PluginPackageManager::new(Config::config_dir());
-                    let result = (|| -> anyhow::Result<PluginManagerOperationOutcome> {
-                        let id = plugin::package::PluginId::parse(&package)?;
-                        let installed = manager
-                            .installed(&id)?
-                            .ok_or_else(|| anyhow::anyhow!("plugin `{id}` is not installed"))?;
-                        manager.trust_package_grammars_exact(&id, &digests)?;
-                        Ok(PluginManagerOperationOutcome {
-                            message: format!(
-                                "Approved the confirmed native grammars for {}.",
-                                installed.name
-                            ),
-                            reload_languages: installed.has_languages,
-                            restart_plugins: false,
-                        })
-                    })();
-                    let (message, reload_languages, restart_plugins) = match result {
-                        Ok(outcome) => (
-                            outcome.message,
-                            outcome.reload_languages,
-                            outcome.restart_plugins,
-                        ),
-                        Err(error) => (
-                            format!("Native grammar approval failed: {error}"),
-                            false,
-                            false,
-                        ),
-                    };
-                    runtime.send_request(PluginRequest::Action(Action::PluginManagerFinished {
-                        message,
-                        reload_languages,
-                        restart_plugins,
-                    }));
-                });
-            }
-            Action::PluginManagerCatalogInstall {
-                catalog_url,
-                package,
-                trust_native_grammars,
-            } => {
-                add_to_history = false;
-                let trust_native_grammars = *trust_native_grammars;
-                let catalog_url = catalog_url.clone();
-                let package = (**package).clone();
-                self.set_legacy_message(Some(format!("Installing {}…", package.name)));
-                let runtime = runtime.clone();
-                tokio::spawn(async move {
-                    let manager = plugin::package::PluginPackageManager::new(Config::config_dir());
-                    let result = manager
-                        .install_catalog_package(&catalog_url, &package, trust_native_grammars)
-                        .await;
-                    let (message, reload_languages) = match result {
-                        Ok(package)
-                            if package.has_native_grammars && !trust_native_grammars => (
-                            format!(
-                                "Installed {} {}. Its native grammar remains disabled until you approve it.",
-                                package.name, package.version
-                            ),
-                            false,
-                        ),
-                        Ok(package) => {
-                            let reload_languages = package.has_languages;
-                            (
-                                format!("Installed {} {}.", package.name, package.version),
-                                reload_languages,
-                            )
-                        }
-                        Err(error) => (format!("Language pack install failed: {error}"), false),
-                    };
-                    runtime.send_request(PluginRequest::Action(Action::PluginManagerFinished {
-                        message,
-                        reload_languages,
-                        restart_plugins: false,
-                    }));
-                });
-            }
-            Action::PluginManagerInstall(source) => {
-                add_to_history = false;
-                let source = source.trim().to_string();
-                if source.is_empty() {
-                    self.set_legacy_message(Some("plugin source cannot be empty".to_string()));
-                } else {
-                    self.set_legacy_message(Some(format!("Installing plugin from {source}…")));
-                    let runtime = runtime.clone();
-                    tokio::spawn(async move {
-                        let manager =
-                            plugin::package::PluginPackageManager::new(Config::config_dir());
-                        let result = if Path::new(&source).exists() {
-                            manager.install_path(Path::new(&source)).await
-                        } else {
-                            let (repository, version) = source
-                                .rsplit_once('@')
-                                .map_or((source.as_str(), None), |(repository, version)| {
-                                    (repository, Some(version))
-                                });
-                            manager.install_github(repository, version).await
-                        };
-                        let (message, reload_languages, restart_plugins) = match result {
-                            Ok(plugin) => (
-                                if plugin.has_native_grammars {
-                                    format!(
-                                        "Installed {} {}. Its native grammar remains disabled until you approve it.",
-                                        plugin.name, plugin.version
-                                    )
-                                } else {
-                                    format!("Installed {} {}.", plugin.name, plugin.version)
-                                },
-                                plugin.has_languages && !plugin.has_native_grammars,
-                                plugin.has_husk || plugin.has_companion,
-                            ),
-                            Err(error) => (format!("Plugin install failed: {error}"), false, false),
-                        };
-                        runtime.send_request(PluginRequest::Action(
-                            Action::PluginManagerFinished {
-                                message,
-                                reload_languages,
-                                restart_plugins,
-                            },
-                        ));
-                    });
-                }
-            }
-            Action::PluginManagerAction(operation) => {
-                add_to_history = false;
-                let operation = operation.clone();
-                self.set_legacy_message(Some("Updating plugin installation…".to_string()));
-                let runtime = runtime.clone();
-                tokio::spawn(async move {
-                    let manager = plugin::package::PluginPackageManager::new(Config::config_dir());
-                    let result = apply_plugin_manager_operation(&manager, &operation).await;
-                    let (message, reload_languages, restart_plugins) = match result {
-                        Ok(outcome) => (
-                            outcome.message,
-                            outcome.reload_languages,
-                            outcome.restart_plugins,
-                        ),
-                        Err(error) => (format!("Plugin operation failed: {error}"), false, false),
-                    };
-                    runtime.send_request(PluginRequest::Action(Action::PluginManagerFinished {
-                        message,
-                        reload_languages,
-                        restart_plugins,
-                    }));
-                });
-            }
-            Action::PluginManagerFinished {
-                message,
-                reload_languages,
-                restart_plugins,
-            } => {
-                add_to_history = false;
-                let mut message = message.clone();
-                if *reload_languages {
-                    match self.reload_languages().await {
-                        Ok(_) => message.push_str(" Language definitions reloaded."),
-                        Err(error) => message.push_str(&format!(
-                            " The package was changed, but language reload failed: {error}"
-                        )),
-                    }
-                }
-                if *restart_plugins {
-                    message.push_str(" Restart Red to refresh plugin code.");
-                }
-                self.set_legacy_message(Some(message.clone()));
-                let manager = plugin::package::PluginPackageManager::new(Config::config_dir());
-                let installed = manager.list().unwrap_or_default();
-                let items = plugin_manager_items(
-                    &installed,
-                    &self.plugin_catalog,
-                    &self.plugin_catalog_url,
-                );
-                self.current_dialog =
-                    Some(Box::new(language_pack_picker(self, items, message, false)));
-                self.render(buffer)?;
+                self.execute_package_action(action, buffer, runtime).await?;
             }
             Action::Command(cmd) => {
                 log!("Handling command: {cmd}");
